Week 6: Project Scenario | House Sales in King County, USA
In this final assignment, I will assume the role of a Data Analyst working at a real estate investment trust organization who wants to start investing in residential real estate. I will be given a dataset containing detailed information about house prices in the region based on a number of property features, and it will be my job to analyze and predict the market price of houses given that information.

This dataset contains house sale prices for King County, which includes Seattle. It includes homes sold between May 2014 and May 2015.